+++ src/main/java/hellfirepvp/modularmachinery/common/integration/crafttweaker/upgrade/DynamicMachineUpgradeBuilder.java
@@ -39,7 +39,7 @@ public class DynamicMachineUpgradeBuilder {
      * @return 构建器
      */
     @ZenMethod
-    public static DynamicMachineUpgradeBuilder newBuilder(String name, String localizedName, int level, int maxStack) {
+    public static DynamicMachineUpgradeBuilder newBuilder(String name, String localizedName, float level, int maxStack) {
         if (RegistryUpgrade.getUpgrade(name) != null) {
             CraftTweakerAPI.logError("[ModularMachinery] Already registered SimpleMachineUpgrade " + name + '!');
             return null;
